model_deploy/requirements_38.txt
glob2==0.7
numpy==1.23.5
opencv-python==4.8.0.76
python-dotenv==1.0.0
tensorflow==2.13.1
# tensorflow-gpu>=2.0 
tensorflow-io==0.31.0
wandb==0.15.12
pydot==1.4.2
# windows: install graphviz for visualize model structure
tqdm==4.66.1

#segmentation part
ultralytics==8.1.14 
segment-anything==1.0

# config file
python-box==7.1.1

# adaptive weight
git+git://github.com/dr-aheydari/SoftAdapt.git